colocate startWatcher with instantiateMaker
