+++ src/providers/MakerProvider.js
@@ -5,6 +5,8 @@ import { mixpanelIdentify } from '../utils/analytics';
 import { instantiateMaker } from '../maker';
 import PropTypes from 'prop-types';
 import { Routes } from 'utils/constants';
+import useStore from '../hooks/useStore';
+import { startWatcher } from '../watch';
 
 export const MakerObjectContext = createContext();
 
@@ -14,6 +16,7 @@ function MakerProvider({ children, network, testchainId, backendEnv }) {
   const [txLastUpdate, setTxLastUpdate] = useState(0);
   const [maker, setMaker] = useState(null);
   const navigation = useNavigation();
+  const [, dispatch] = useStore();
 
   const initAccount = account => {
     mixpanelIdentify(account.address, 'metamask');
@@ -21,14 +24,24 @@ function MakerProvider({ children, network, testchainId, backendEnv }) {
   };
 
   useEffect(() => {
-    instantiateMaker({
-      network,
-      testchainId,
-      backendEnv
-    }).then(maker => {
-      if (maker.service('accounts').hasAccount())
+    (async () => {
+      const maker = await instantiateMaker({
+        network,
+        testchainId,
+        backendEnv
+      });
+      if (maker.service('accounts').hasAccount()) {
         initAccount(maker.currentAccount());
+      }
       setMaker(maker);
+
+      const scs = maker.service('smartContract');
+      startWatcher({
+        rpcUrl: maker.service('web3').rpcUrl,
+        multicallAddress: scs.getContractAddress('MULTICALL'),
+        addresses: scs.getContractAddresses(),
+        dispatch
+      });
 
       maker.on('accounts/CHANGE', eventObj => {
         const { account } = eventObj.payload;
@@ -45,8 +58,8 @@ function MakerProvider({ children, network, testchainId, backendEnv }) {
           }
         })();
       });
-    });
-  }, [backendEnv, network, testchainId]);
+    })();
+  }, [backendEnv, dispatch, network, testchainId]);
 
   const checkForNewCdps = async (numTries = 5, timeout = 500) => {
     const proxy = await maker.service('proxy').getProxyAddress(account.address);
